Fixes #3
ONERR now puts the labe l in the output stream so it cab be searched for when generating
QCode. A fix-up scheme is maybe best as it would be nice to be able to generate QCode as the
intermedite code arrives. Tat reduces the amount of memory needed if the intermediate buffers aren't needed.
Labels are now inthe output stream too.